requirements.txt
# 基础依赖
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
sqlalchemy>=2.0.22
alembic>=1.12.0
python-dotenv>=1.0.0
# 文档处理
pypdf>=3.15.1
python-docx>=0.8.11
openpyxl>=3.1.2
pytesseract>=0.3.10
pillow>=10.0.1
# 向量化与搜索
transformers>=4.34.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.1
# API与Web
jinja2>=3.1.2
aiofiles>=23.2.1
python-multipart>=0.0.6
# 测试
pytest>=7.4.2
httpx>=0.25.0
pandas~=2.2.3
chardet~=5.2.0
passlib~=1.7.4
jieba~=0.42.1
numpy~=2.2.3
setuptools~=76.0.0
sympy~=1.13.1
torch~=2.6.0
jwt~=1.3.1
PyJWT>=2.0.0